pyproject.toml
[project]
name = "uscre-login"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "flask>=3.1.2",
    "flask-mail>=0.10.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "loguru>=0.7.3",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
]
